+++ test/bitcrowd_ecto/tabular_test.exs
@@ -1,0 +1,5 @@
+defmodule BitcrowdEcto.TabularTest do
+  use ExUnit.Case, async: true
+  import BitcrowdEcto.Tabular
+  doctest BitcrowdEcto.Tabular
+end